+++ app/build.gradle.kts
@@ -77,9 +77,9 @@ dependencies {
     implementation(libs.javapoet)
 
     implementation ("androidx.viewpager2:viewpager2:1.1.0")
-    implementation ("com.google.android.material:material:1.11.0")
+    implementation ("androidx.navigation:navigation-fragment-ktx:2.7.6")
+    implementation ("androidx.navigation:navigation-ui-ktx:2.7.6")
 
-    // ИЛИ: implementation("com.squareup:javapoet:1.13.0")
 
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
