+++ app/src/main/AndroidManifest.xml
@@ -18,10 +18,10 @@
         tools:targetApi="31">
         <activity
             android:name=".main.ui.MainActivity"
-            android:exported="true">
+            android:exported="true"
+            android:windowSoftInputMode="adjustResize">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
 
@@ -29,11 +29,5 @@
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
-        <activity android:name=".settings.ui.SettingsActivity" />
-        <activity android:name=".search.ui.SearchActivity" />
-        <activity android:name=".mediateka.MediatekaActivity" />
-        <activity android:name=".player.ui.AudioPlayerActivity" />
-
     </application>
-
 </manifest>
