+++ app/src/main/java/com/practicum/playlistmaker/main/ui/MainActivity.kt
@@ -1,33 +1,62 @@
 package com.practicum.playlistmaker.main.ui
 
-import android.content.Intent
 import android.os.Bundle
+import android.util.Log
 import android.view.View
+import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.NavController
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.practicum.playlistmaker.R
-import com.practicum.playlistmaker.mediateka.MediatekaActivity
-import com.practicum.playlistmaker.search.ui.SearchActivity
-import com.practicum.playlistmaker.settings.ui.SettingsActivity
 
 class MainActivity : AppCompatActivity() {
+
+    private lateinit var navController: NavController
+    private lateinit var bottomNavigationView: BottomNavigationView
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        findViewById<View>(R.id.setting).setOnClickListener {
-            val intent = Intent(this, SettingsActivity::class.java)
-            startActivity(intent)
+        try {
+            val navHostFragment = supportFragmentManager
+                .findFragmentById(R.id.nav_host_fragment) as NavHostFragment
+            navController = navHostFragment.navController
+            bottomNavigationView = findViewById(R.id.bottom_navigation)
+            bottomNavigationView.setupWithNavController(navController)
+            setupNavigation()
+            Log.d("MainActivity", "Navigation setup completed successfully")
+        } catch (e: Exception) {
+            Log.e("MainActivity", "Error initializing navigation", e)
+            Toast.makeText(
+                this,
+                "Ошибка навигации: ${e.message}",
+                Toast.LENGTH_LONG
+            ).show()
         }
+    }
 
-        findViewById<View>(R.id.search_button).setOnClickListener {
-            val intent = Intent(this, SearchActivity::class.java)
-            startActivity(intent)
+    private fun setupNavigation() {
+        navController.addOnDestinationChangedListener { _, destination, _ ->
+            when (destination.id) {
+                R.id.audioPlayerFragment -> {
+                    bottomNavigationView.visibility = View.GONE
+                }
+                else -> {
+                    bottomNavigationView.visibility = View.VISIBLE
+                }
+            }
         }
+    }
 
-        findViewById<View>(R.id.mediateca).setOnClickListener {
-            val intent = Intent(this, MediatekaActivity::class.java)
-            startActivity(intent)
+    override fun onBackPressed() {
+        val currentDestinationId = navController.currentDestination?.id
+        if (currentDestinationId == R.id.mediatekaFragment) {
+            finish()
+        } else {
+            super.onBackPressed()
         }
     }
 }
