+++ app/src/main/java/com/practicum/playlistmaker/mediateka/ui/adapter/MediatekaViewPagerAdapter.kt
@@ -2,13 +2,16 @@ package com.practicum.playlistmaker.mediateka.ui.adapter
 
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.practicum.playlistmaker.mediateka.ui.fragment.FragmentFavorites
 import com.practicum.playlistmaker.mediateka.ui.fragment.FragmentPlaylists
 
 class MediatekaViewPagerAdapter(
-    private val activity: AppCompatActivity
-) : FragmentStateAdapter(activity) {
+    fragmentManager: FragmentManager,
+    lifecycle: Lifecycle
+) : FragmentStateAdapter(fragmentManager, lifecycle) {
 
     override fun getItemCount(): Int = 2
