+++ app/src/main/res/color/bottom_nav_colors.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+    <selector xmlns:android="http://schemas.android.com/apk/res/android">
+        <!-- Цвет для активного пункта (YP Blue) -->
+        <item android:color="@color/yp_blue" android:state_checked="true" />
+
+        <!-- Цвет для неактивных пунктов (YP Black) -->
+        <item android:color="@color/yp_icon" />
+    </selector>
+
