+++ app/src/main/res/layout/fragment_audio_player.xml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+    <TextView
+        android:id="@+id/back"
+        style="@style/TextStyleTopScreen"
+        android:text="@string/back"
+        android:layout_marginBottom="26dp" />
+
+    <LinearLayout android:id="@+id/no_results_layout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:paddingTop="86dp"
+        android:gravity="center"
+        android:visibility="gone"
+        xmlns:android="http://schemas.android.com/apk/res/android">
+
+        <ImageView
+            android:id="@+id/no_results_image"
+            style="@style/NoResultateImageSearch"/>
+
+        <TextView
+            android:id="@+id/no_results_text"
+            style="@style/ErrorSearch"
+            android:text="@string/no_results"/>
+
+    </LinearLayout>
+
+    <!-- Список результатов поиска -->
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/recyclerViewAudioPlayer"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager" />
+
+
+</LinearLayout>
