+++ app/src/main/res/layout/fragment_search.xml
@@ -1,0 +1,143 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <!-- Основной контент экрана (был корневым LinearLayout) -->
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        style="@style/BackgroundSearch">
+
+        <!-- Заголовок или кнопка возврата -->
+        <TextView
+            android:id="@+id/back"
+            style="@style/TextStyleTopScreen"
+            android:text="@string/search"
+            android:layout_marginBottom="8dp"/>
+
+        <!-- Поле поиска с кнопкой сброса -->
+        <FrameLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="24dp">
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="vertical">
+
+                <EditText
+                    android:id="@+id/search_edit_text"
+                    style="@style/TextStyleSearch"
+                    android:drawableStart="@drawable/ic_search_16"
+                    android:hint="@string/search"
+                    android:imeOptions="actionDone"
+                    android:singleLine="true"
+                    android:paddingEnd="40dp" />
+
+                <!--                <TextView-->
+                <!--                    android:id="@+id/searchHint"-->
+                <!--                    android:padding="16dp"-->
+                <!--                    style="@style/SearchHint"-->
+                <!--                    android:text="@string/search_hint"-->
+                <!--                    android:visibility="gone" />-->
+            </LinearLayout>
+
+            <ImageView
+                android:id="@+id/reset_button"
+                style="@style/ResetStyleButton"
+                android:layout_gravity="end|center_vertical"
+                android:src="@drawable/ic_delete_vector_12" />
+        </FrameLayout>
+
+        <!-- Сообщение "Нет результатов" -->
+        <LinearLayout
+            android:id="@+id/no_results_layout"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical"
+            android:paddingTop="86dp"
+            android:gravity="center"
+            android:visibility="gone">
+
+            <ImageView
+                android:id="@+id/no_results_image"
+                style="@style/NoResultateImageSearch" />
+
+            <TextView
+                android:id="@+id/no_results_text"
+                style="@style/ErrorSearch"
+                android:text="@string/no_results" />
+        </LinearLayout>
+
+        <!-- Сообщение об ошибке -->
+        <LinearLayout
+            android:id="@+id/error_layout"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical"
+            android:paddingTop="86dp"
+            android:gravity="center"
+            android:visibility="gone">
+
+            <ImageView
+                android:id="@+id/error_image"
+                style="@style/ErrorImageSearch" />
+
+            <TextView
+                android:id="@+id/error_text"
+                style="@style/ErrorSearch"
+                android:text="@string/error_occurred1" />
+
+            <Button
+                android:id="@+id/refresh_button"
+                style="@style/ButtonStyleErrorSearch"
+                android:text="@string/update"
+                android:enabled="true" />
+        </LinearLayout>
+
+        <!-- История поиска -->
+        <LinearLayout
+            android:id="@+id/search_history_layout"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical"
+            android:visibility="gone">
+
+            <TextView
+                android:id="@+id/history_title"
+                android:text="@string/you_were_looking_for"
+                style="@style/HistorySearch" />
+
+            <androidx.recyclerview.widget.RecyclerView
+                android:id="@+id/history_recycler_view"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager" />
+
+            <Button
+                android:id="@+id/clear_history_button"
+                android:text="@string/clear_the_history"
+                style="@style/ButtonHistorySearch" />
+        </LinearLayout>
+
+        <!-- Список результатов поиска -->
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/recyclerView"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager" />
+    </LinearLayout>
+
+    <ProgressBar
+        android:id="@+id/progressBar"
+        android:layout_width="44dp"
+        android:layout_height="44dp"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginTop="248dp"
+        android:indeterminateTint="@color/progressbar_blue"
+        android:visibility="gone" />
+</FrameLayout>
