+++ app/src/main/res/layout/fragment_settings.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    style="@style/BackgroundSetting"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+    <TextView
+        android:id="@+id/back"
+        style="@style/TextStyleTopScreen"
+        android:text="@string/setting" />
+
+    <com.google.android.material.switchmaterial.SwitchMaterial
+        android:id="@+id/switch_button"
+        style="@style/TextStyleSetting"
+        android:text="@string/black_theme" />
+
+    <TextView
+        android:id="@+id/btnShare"
+        style="@style/TextStyleSetting"
+        android:text="@string/share_the_app"
+        app:drawableEndCompat="@drawable/ic_share_24" />
+
+
+    <TextView
+        android:id="@+id/supportButton"
+        style="@style/TextStyleSetting"
+        android:text="@string/write_to_support"
+        app:drawableEndCompat="@drawable/ic_support_24" />
+
+    <TextView
+        android:id="@+id/userAgreementButton"
+        style="@style/TextStyleSetting"
+        android:text="@string/user_agreement"
+        app:drawableEndCompat="@drawable/ic_light_mode_24" />
+
+</LinearLayout>
+
+
