+++ app/src/main/res/menu/menu_bottom_navigation.xml
@@ -1,0 +1,14 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/searchFragment"
+        android:icon="@drawable/ic_search_16"
+        android:title="@string/search_button" />
+    <item
+        android:id="@+id/mediatekaFragment"
+        android:icon="@drawable/ic_mediateca_24"
+        android:title="@string/mediateca" />
+    <item
+        android:id="@+id/settingsFragment"
+        android:icon="@drawable/ic_setting_24"
+        android:title="@string/setting" />
+</menu>
