+++ app/src/main/res/values-night/colors_night.xml
@@ -16,4 +16,6 @@
     <color name="yp_light_gray">#E6E8EB</color>
     <color name="yp_text_gray">#AEAFB4</color>
     <color name="for_point_search">#FFFFFF</color>
+    <color name="yp_icon">#FFFFFFFF</color>
+
 </resources>
