+++ app/src/main/res/values/colors.xml
@@ -19,6 +19,7 @@
     <color name="without_black">#00000000</color>
     <color name="for_point_search">#AEAFB4</color>
     <color name="progressbar_blue">#3772E7</color>
+    <color name="yp_icon">#1A1B22</color>
 
 
 </resources>
