+++ app/src/main/res/values/strings.xml
@@ -33,6 +33,8 @@
     <string name="new_playlist">Новый плейлист</string>
     <string name="no_playlists_message">Вы не создали\nни одного плейлиста</string>
     <string name="empty_state_image">Изображение состояния отсутствия данных</string>
+    <string name="audio_player">Аудиоплеер</string>
+
