+++ app/src/main/res/values/styles.xml
@@ -95,6 +95,7 @@
         <item name="android:drawableTint">@color/yp_black</item>
         <item name="android:textFontWeight">500</item>
         <item name="android:drawablePadding">12dp</item>
+        <item name="android:paddingStart">12dp</item>
 
     </style>
